requirements.txt
pafy
youtube-dl
pyyaml
